+++ frontend/src/sections/history/Standardization/Standardization.tsx
@@ -1,0 +1,73 @@
+/**
+ *Salamaik Client 
+ * © Kanwil DJPb Sumbar 2024
+ */
+
+import { useEffect, useState } from "react";
+import StandardizationTable from "./components/StandardizationTable";
+import useStandardization from "./useStandardization";
+import usePreviewFileModal from "./usePreviewFileModal";
+import PreviewFileModal from "./components/PreviewFileModal";
+import { Stack } from "@mui/material";
+import PageLoading from "../../../components/pageLoading";
+// --------------------------------------------------------------
+
+interface StandardizationProps{
+  selectedUnit: string,
+  selectedPeriod: number,
+  selectedData: number
+}
+// --------------------------------------------------------------
+export default function Standardization({selectedUnit, selectedPeriod, selectedData}:StandardizationProps) {
+  const [loading, setLoading] = useState(true);
+
+  const {open, modalOpen, modalClose, file} = usePreviewFileModal();
+
+  const {getStandardization} = useStandardization();
+
+  useEffect(() => {
+    async function getData(){
+      try{
+        await getStandardization(selectedUnit, selectedPeriod);
+        setLoading(false);
+      }catch(err){
+        setLoading(false);
+      }
+    }
+
+    setLoading(true);
+    getData();
+
+  }, [selectedUnit, selectedPeriod, selectedData]);
+
+  return (
+    <>
+       <Stack direction='column' spacing={4}>
+          {loading
+            ?<PageLoading duration={2}/>
+            : <>
+                <StandardizationTable 
+                  header={'Manajemen Eksternal'} 
+                  modalOpen={modalOpen} 
+                  kppnTab={selectedUnit}
+                  cluster={1} 
+                />
+                <StandardizationTable 
+                  header={'Penguatan Kapasitas Perbendaharaan'} 
+                  modalOpen={modalOpen} 
+                  kppnTab={selectedUnit} 
+                  cluster={2} 
+                />
+                <StandardizationTable 
+                  header={'Penguatan Manajemen Internal'} 
+                  modalOpen={modalOpen} 
+                  kppnTab={selectedUnit}
+                  cluster={3}  
+                />
+              </>
+          }
+        </Stack>
+        <PreviewFileModal open={open} modalClose={modalClose} file={file} kppnId={selectedUnit} />
+    </>
+  )
+}